Fix various compiler warnings
- Remove unused variables and class members in conv generator
- Change several if/else if/... blocks to switch/case to remove
- warnings about potentially uninitialized variables and to give us a
- warning if we add a case in the future
- Turn some can't happen cases into llvm_unreachable() so they crash
- the compiler if they do, in fact, happen, instead of plowing ahead
- with an error message
- Fix random utilities by:
  - Using an abs() that won't truncate
  - Performing certain calculations with doubles since RAND_MAX is not
    re presentable by a float on 64-bit platforms

The remaining pile of warnings that's our responsibility is in
rocm-runtime-wrappers.cpp , which has extern "C" functions that export
templated types that probably shouldn't be extern "C" (the non 1D
StridedMemRefType<>s) - it's not clear what to do about those.

+++ mlir/lib/ExecutionEngine/ROCm/rocm-runtime-wrappers.cpp
@@ -390,7 +390,8 @@ short randomIntegerValue(short min, short max) {
 float randomFloatValue(short min, short max) {
   if (min == max)
     return (float)min;
-  return (max - min) * (float)(std::rand()) / RAND_MAX + (float)min;
+  return (float)((max - min) * (double)(std::rand()) / (double)RAND_MAX) +
+         (float)min;
 }
 
 // 3D float memref utility routines.
